fix: index expression mocking
Before when we mocked index expressions, we would get an 'Invalid index' diagnostic, since we'd mock the key to value that was never in the collection.

This updates it to first check if the mocking causes an 'Invalid index' diagnostic, and if it does we return the first element from the collection. This may be the incorrect value, but seems the safest since at least the value returned will be the expected type.

If we see other diagnostics we return an empty string value. This may cause issues if the collection isn't meant to be a list of strings, but since we can't evaluate it this seems the safest.