fix: add support for `task_definitions` referenced through `ecs_task_set`. (#3078)
Adds support to fetch the associated `ecs_service` task definition by walking references
through a `ecs_task_set` resource. Referencing task defitions through a task set is common
and without support for this we show 0 costs on a correctly configured FARGATE `ecs_service`.

An ECS task set is described by AWS as group of tasks within an Amazon ECS service
that enables the management and deployment of multiple service revisions. Notably, it is
a common way to use load balancers with an ecs service.