fix: panic when cloud run resources do not have usage (#3080)
Resolves issues where the newly added cloud run resources panic when they
do not have any usage defined. I've reworked some of the methods in the various
classes to handle these cases. I've also removed the old `CloudRunService` and
`CloudRunV2Service` and merged them into one resource. They were identical appart
from the input parameters and the fact that that one had a different cost component
description "always-on", which I don't think is an issue.