fix: dependency_paths detection issues
Fixes two issues with the `dependency_paths` logic

* Adds wildcard suffix to the generated `dependency_paths` for autodetect. This
is done so that changes in nested files are detected on Infracost's hosted services.
* Merges the tfvars into the `DependencyPaths` method, this resolves an issue where
if no module calls were detected then only tfvars would be added to the `dependency_paths`
config key. This would omit the project from the changes and therefore only tfvar changes
would cause a rerun on hosted services. This also standardises things for those users
who use a template, as now the tfvars are part of the `.DependencyPaths` project template
variable.