refactor: add `region` field to resource data
Adds a `Region` property to the `schema.ResourceData` which is set when
parsing the plan and set before evaluating each individual resource. This
property is meant to replace the old `d.Get` approach which sets the region
directly on the `RawValues` of the json.

This new approach sets the `Region` after the references have been parsed and
changes the `azure` region functionality to use the `getLocation` logic that was
previously done on demand per resource.