refactor: add `GetRegion` func to `RegistryItem`
Adds a resource specific `GetRegion` property to the `RegistryItem` which
can be used to override the provider default functionality for fetching a
region for a resource.

Changes usage of `lookupRegion` functions in azure resources to use the new
`ResourceData.Region` property which is computed before evaluation.

Changes the `ResourceRegistryMap` to no longer use a `sync.Once` and instead be
built as a singleton `var`. This removes the sync requirement as callers only read
from this map.