refactor: add `region` field to resource data (#3107)
* refactor: add `region` field to resource data

Adds a `Region` property to the `schema.ResourceData` which is set when
parsing the plan and set before evaluating each individual resource. This
property is meant to replace the old `d.Get` approach which sets the region
directly on the `RawValues` of the json.

This new approach sets the `Region` after the references have been parsed and
changes the `azure` region functionality to use the `getLocation` logic that was
previously done on demand per resource.

* refactor: add `GetRegion` func to `RegistryItem`

Adds a resource specific `GetRegion` property to the `RegistryItem` which
can be used to override the provider default functionality for fetching a
region for a resource.

Changes usage of `lookupRegion` functions in azure resources to use the new
`ResourceData.Region` property which is computed before evaluation.

Changes the `ResourceRegistryMap` to no longer use a `sync.Once` and instead be
built as a singleton `var`. This removes the sync requirement as callers only read
from this map.

* fix: recursively set the reference region if they are not already set