refactor: use global logger
This change removes the use of child loggers with additional metadata. Additionally
it changes log levels for a number of calls from Debug to Trace. These changes have
been motivated for the following reasons:

1. Performance - investigation on certain large repos has identified that the `With`
   functionality to create child loggers is causing an alarming amount of allocations.
2. Debugging - The previous logging logic with attributes and metadata crammed into the
   `Debug` level was crazy to debug. In reality much of these lines should be `Trace`,
   and the additional attributes were not useful for filtering.

This change should simplify logging across the codebase, speed up some repos and aid in debugging.