feat: add external module metadata to resources (#3130)
Adds additional metadata to Infracost resources which are used in external modules.
This adds a fully qualified external URL to the remote where the resource lives, referencing
the file and version used. Additionally this PR adds helper metadata to flatten the first
module call into the resource metadata so it is easier for the dashboard to work with.