fix: errored projects should show correct project name suffix
Errored projects weren't showing the `-<env>` suffix on their project names in the JSON. This causes `found duplicate project name` issues when running `infracost diff`.
This fixes it by making sure that the module always has the suffix, so that if it errors and returns early we still know the suffix. Previously this was only added if the module was parsed successfully.