enhance: don't error on invalid YAML and JSON
We don't want to stop the whole config template execution on this. It seems safer to log and not panic on these errors, since we shouldn't fail runs if we're unable to parse them.

In the future if we want an option that does return error we can do what sprig does and create `mustParseYaml` and `mustParseJson` functions.