fix: adds module path to errored project
Adds `TerraformModulePath` to any errored project. This resolves an issue in the `diff`
command. Errored projects that were discovered at a child path of the initial path
provided in the config would have the same `Path` and or `Name` metadata. This would
cause a `compare-to` conflict because we saw them as duplicate projects, even though these
projects were actually located in different directories.