fix: terraform env matching on colon separated filenames (#3170)
Resolves issue where a terraform var file that started with an env name but was
preceded by a `.` would fail to match an env, e.g `dev.foo.tfvars`. This was because
of a typo in the prefix matching logic. Changing this logic also exposed an issue when
matching hidden file var files (those starting with a `.`). As the previous typo was
allowing for them to be successfully matched. Now that I've fixed the prefix matching
logic, i've also changed the `clean` logic to remove `.` from the filenames so that
the hidden files -> env name matching works as expected.