feat: add spacelift remote variable support
Adds a remote variable loader which will fetch defined environment contexts for
a stack matching the project repository and env name. In order to accomodate this
I've changed the `hcl.Parser` var loader properties to a slice, this way we can
define multiple remote loaders in response to a users configuration.