refactor(autodect): exclude examples, include hidden dirs
Changes autodetect logic to exclude `examples` folders from autodetect. This is a
common pattern for repos that are used as modules. Including the `examples` projects
has lead to confusion and false positives with cost changes/policy failures. Therefore
it has been decided that we should exclude these by default.

Additionally, autodetect logic has been altered to include project from hidden folders.
The initial logic (where we excluded hidden directories) was made as we falsely assumed
users would not place valid Terraform in hidden directories. This has been proven wrong,
so now we only exclude a handful of known hidden dirs (e.g. `.infracost`) from the project
detection logic.

I've also made a change to the `shouldUseProject` logic in the project detection algorithm so
that users can use the `include_dirs` logic to override any of the changes above. E.g. if
a user wanted to include the `examples` directory for a repository.