fix: cleanup bad module download directories
We need to clean up any partially downloaded modules if there is an error (e.g. if the module was downloaded but the ref couldn't be checked out), otherwise when another project tries to download the same module the `_, err = os.Stat(dest)` will decide that it has been previously downloaded successfully.