fix: duplicate projects with force with invalid project
Resolves issue where invalid terraform project (no backend/provider blocks)
are found in the same directory as a terragrunt project. If we try and force
the project type in this scenario, the project locator resolve 0 valid projects
and then fell back to just adding in all projects. I've fixed this by only "forcing"
the fallback projects if they are of the forced project type.