fix: treat Terraform module errors as Terragrunt project errors (#3320)
Prior to this change, when evaluating a Terragrunt project, we treated the
underlying Terraform module errors as just a warning diagnostic. This created
issues with the diff logic as baseline Terragrunt projects would have warning
metadatas but a zero cost as the project could not be evaluated. If this was an
intermittent error, which is often the case with Terragrunt and the caching logic
then users would see a 100% cost increase for the project as it was seen as going
from 0 to the current project price.

I've changed the logic now to treat Terraform errors as first class Terragrunt evaluation
errors in the evaluator. This means that when diffing we'll show that we couldn't compute
the price instead of a new project addition.