fix: Add local modules to manifest if they are source mapped
When module's source has a relative path like `../../modules/` but it
needs to be mapped to another path, the module is still loaded using
original source value.
Important: the source map will only work when the path starts with the
source map key. It won't replace it in the middle of the path.