fix(output): add nil check for Project.PastBreakdown to output.Load
The code in internal/output implicitly assumes a Project's PastBreakdown
field is non-nil (though the Breakdown fields *may* be nil / empty).

This commit adds a check to the json loader to ensure that if the
PastBreakdown is nil in the unmarshaled struct, we set it to a non-nil
(but otherwise empty) Breakdown.